refactor: point 패키지 이동 및 메소드명 수정
