+++ apps/commerce-api/src/main/java/com/loopers/application/provided/PointFinder.java
@@ -1,7 +1,7 @@
 package com.loopers.application.provided;
 
 import com.loopers.domain.member.MemberId;
-import com.loopers.domain.member.Point;
+import com.loopers.domain.member.point.Point;
 
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.NotNull;
